Added changes

+++ final_work/temp_function.h
@@ -38,11 +38,11 @@ void add_record(sensor info[], int number, uint16_t year, uint8_t month,
                 uint8_t day, uint8_t hour, uint8_t minute, int8_t temp);
 int read_data_file(FILE **fptr, sensor info[]);
 double average_month_temp(sensor info[], int number, int month);
-int min_month_temp(sensor info[], int number, int month);
-int max_month_temp(sensor info[], int number, int month);
+int8_t min_month_temp(sensor info[], int number, int month);
+int8_t max_month_temp(sensor info[], int number, int month);
 double average_year_temp(sensor info[], int number);
-int min_year_temp(sensor info[], int number);
-int max_year_temp(sensor info[], int number);
+int8_t min_year_temp(sensor info[], int number);
+int8_t max_year_temp(sensor info[], int number);
 void year_statistic(sensor info[], int number);
 void month_statistic(sensor info[], int number, uint8_t month_number);
 void num_month_row(int month);
